Account dropdown for header fixing nav issues as well
